:sparkles:  feat: 내가 작성한 게시물이 아닌 경우 수정 및 삭제 불가능하도록 수정

+++ src/main/kotlin/com/kotlin/study/dongambackend/domain/post/service/PostService.kt
@@ -1,6 +1,8 @@
 package com.kotlin.study.dongambackend.domain.post.service
 
+import com.kotlin.study.dongambackend.common.exception.common.ForbiddenException
 import com.kotlin.study.dongambackend.common.exception.common.NotFoundException
+import com.kotlin.study.dongambackend.common.type.ResponseStatusType
 import com.kotlin.study.dongambackend.domain.post.validator.type.BoardCategory
 import com.kotlin.study.dongambackend.domain.post.entity.id.PostLikeId
 import com.kotlin.study.dongambackend.domain.post.dto.request.PostCreateRequest
@@ -51,15 +53,24 @@ class PostService(
         return postRepository.save(post).id
     }
 
-    fun updatePost(postUpdateRequest: PostUpdateRequest, postId: Long) {
+    fun updatePost(postUpdateRequest: PostUpdateRequest, postId: Long, userId: Long? = 0L) {
         val post = this.getPost(postId)
+
+        // TODO: 인터셉터 추가
+        if (post.userId.id != userId) {
+            throw ForbiddenException(ResponseStatusType.POST_FORBIDDEN)
+        }
 
         post.updatePost(postUpdateRequest)
         postRepository.save(post)
     }
 
-    fun deletePost(postId: Long) {
-        this.getPost(postId)
+    fun deletePost(postId: Long, userId: Long? = 0L) {
+        val post = this.getPost(postId)
+
+        if (post.userId.id != userId) {
+            throw ForbiddenException(ResponseStatusType.POST_FORBIDDEN)
+        }
 
         postRepository.deleteById(postId)
     }
